Started working on the frontend

+++ backend/controllers/product.controller.js
@@ -32,12 +32,17 @@ export const createProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: `Product ID: ${id} not found` });
+  }
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({ success: true, message: "Product deleted" });
   } catch (error) {
     console.log("Error Deleting Product", error.message);
-    res.status(404).json({ success: false, message: "product not found." });
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
